Add Batt Status to CSV, Update README, CSV Order
I forgot to added the Battery Status to the CSV in the previous commit.
I added it before EstimatedChargeRemaining because that field is
handy to have next to Timestamp for easy charting.
I also forgot to update the Readme with parameter details in the
previous commit, so that has been added.
Finally, I also forgot to mention that I changed the order of the
DesignedCapacity and FullChargeCapacity fields in the previous
commit, so I'm mentioning it here. It makes more sense to have
the columns now in this order. Also added a sanity check to ensure
that PollingFrequency does not go below 1.